+++ wp-content/plugins/creative-mail-by-constant-contact/src/views/admin-feedback-notice/few-contacts.php
@@ -6,6 +6,6 @@
         </p>
         <p><?= __( 'These contacts are already in Creative Mail, send a quick campaign...', 'ce4wp' ); ?></p>
     </section>
-    <button class="button button-primary" onclick="ce4wpNavigateToDashboard(this, 'd25f690a-217a-4d68-9c58-8693965d4673', undefined, ce4wpWidgetStartCallback, ce4wpWidgetFinishCallback)"><?= __( 'Get started', 'ce4wp' ); ?></button>
+    <button class="button button-primary" onclick="ce4wpNavigateToDashboard(this, 'd25f690a-217a-4d68-9c58-8693965d4673', { source: 'feedback_notice' }, ce4wpWidgetStartCallback, ce4wpWidgetFinishCallback)"><?= __( 'Get started', 'ce4wp' ); ?></button>
     <span id="close" onclick="hideAdminFeedbackNotice('feedback_notice_few_contacts')"></span>
 </div>