+++ wp-content/plugins/creative-mail-by-constant-contact/src/views/admin-feedback-notice/many-contacts.php
@@ -10,6 +10,6 @@
             <?= __( 'contacts are ready for a Creative Mail email campaign. Send one now!', 'ce4wp' ); ?>
         </p>
     </section>
-    <button class="button button-primary" onclick="ce4wpNavigateToDashboard(this, 'd25f690a-217a-4d68-9c58-8693965d4673', undefined, ce4wpWidgetStartCallback, ce4wpWidgetFinishCallback)"><?= __( 'Get started', 'ce4wp' ); ?></button>
+    <button class="button button-primary" onclick="ce4wpNavigateToDashboard(this, 'd25f690a-217a-4d68-9c58-8693965d4673', { source: 'feedback_notice' }, ce4wpWidgetStartCallback, ce4wpWidgetFinishCallback)"><?= __( 'Get started', 'ce4wp' ); ?></button>
     <span id="close" onclick="hideAdminFeedbackNotice('feedback_notice_many_contacts')"></span>
 </div>